Add unit tests for BrandingService configuration functionality
